Include SMS auth form in SMS namespace
Rename and move the form.
Renamed App form to match the SMS one, but didn't namespace it since
APP namespace would only have this class.
