+++ cosmetics-web/app/forms/secondary_authentication/app_auth_form.rb
@@ -1,5 +1,5 @@
 module SecondaryAuthentication
-  class AppForm
+  class AppAuthForm
     include ActiveModel::Model
     include ActiveModel::Attributes
